feat: endpoint/use case implementation & other changes
- Feature: Functional endpoint implementation of use case
- Change: Renamed interfaces to "I..." format
- Change: id types from Integer to Long
- Change: return types from endpoint layers to Optional for simplifying handling
- Feature: added mapping function in Request/Response objects
- Change: updated tests to reflect the ID type fix

+++ src/test/java/com/matias/app/test/infrastructure/in/rest/provider/ProductPriceByDateArgumentsProvider.java
@@ -16,8 +16,8 @@ import com.matias.app.infrastructure.adapter.in.rest.response.ProductPriceByDate
 	
 public class ProductPriceByDateArgumentsProvider implements ArgumentsProvider {
 
-	final static Integer BRAND_TEST_ID = 1;
-	final static Integer PRODUCT_TEST_ID = 35455;
+	final static Long BRAND_TEST_ID = Long.valueOf(1);
+	final static Long PRODUCT_TEST_ID = Long.valueOf(35455);
 	
 	// This list contains all requests that are going to be tested
 	final static List<ProductPriceByDateRequest> REQUEST_TEST_LIST = Stream.of(
@@ -30,11 +30,11 @@ public class ProductPriceByDateArgumentsProvider implements ArgumentsProvider {
 	
 	// This list contains the response for each request in the list above (in order)
 	final static List<ProductPriceByDateResponse> RESPONSE_TEST_LIST = Stream.of(
-			ProductPriceByDateResponse.builder().brandId(BRAND_TEST_ID).productId(PRODUCT_TEST_ID).startDate(shortParse("2020-06-14T00:00:00")).endDate(shortParse("2020-12-31T23:59:59")).priceList(Integer.valueOf(1)).price(Double.valueOf(35.50)).build(),
-			ProductPriceByDateResponse.builder().brandId(BRAND_TEST_ID).productId(PRODUCT_TEST_ID).startDate(shortParse("2020-06-14T15:00:00")).endDate(shortParse("2020-06-14T18:30:00")).priceList(Integer.valueOf(2)).price(Double.valueOf(25.45)).build(),
-			ProductPriceByDateResponse.builder().brandId(BRAND_TEST_ID).productId(PRODUCT_TEST_ID).startDate(shortParse("2020-06-14T00:00:00")).endDate(shortParse("2020-12-31T23:59:59")).priceList(Integer.valueOf(1)).price(Double.valueOf(35.50)).build(),
-			ProductPriceByDateResponse.builder().brandId(BRAND_TEST_ID).productId(PRODUCT_TEST_ID).startDate(shortParse("2020-06-15T00:00:00")).endDate(shortParse("2020-06-15T11:00:00")).priceList(Integer.valueOf(3)).price(Double.valueOf(30.50)).build(),
-			ProductPriceByDateResponse.builder().brandId(BRAND_TEST_ID).productId(PRODUCT_TEST_ID).startDate(shortParse("2020-06-15T16:00:00")).endDate(shortParse("2020-12-31T23:59:59")).priceList(Integer.valueOf(4)).price(Double.valueOf(38.95)).build())
+			ProductPriceByDateResponse.builder().brandId(BRAND_TEST_ID).productId(PRODUCT_TEST_ID).startDate(shortParse("2020-06-14T00:00:00")).endDate(shortParse("2020-12-31T23:59:59")).priceList(Long.valueOf(1)).price(Double.valueOf(35.50)).build(),
+			ProductPriceByDateResponse.builder().brandId(BRAND_TEST_ID).productId(PRODUCT_TEST_ID).startDate(shortParse("2020-06-14T15:00:00")).endDate(shortParse("2020-06-14T18:30:00")).priceList(Long.valueOf(2)).price(Double.valueOf(25.45)).build(),
+			ProductPriceByDateResponse.builder().brandId(BRAND_TEST_ID).productId(PRODUCT_TEST_ID).startDate(shortParse("2020-06-14T00:00:00")).endDate(shortParse("2020-12-31T23:59:59")).priceList(Long.valueOf(1)).price(Double.valueOf(35.50)).build(),
+			ProductPriceByDateResponse.builder().brandId(BRAND_TEST_ID).productId(PRODUCT_TEST_ID).startDate(shortParse("2020-06-15T00:00:00")).endDate(shortParse("2020-06-15T11:00:00")).priceList(Long.valueOf(3)).price(Double.valueOf(30.50)).build(),
+			ProductPriceByDateResponse.builder().brandId(BRAND_TEST_ID).productId(PRODUCT_TEST_ID).startDate(shortParse("2020-06-15T16:00:00")).endDate(shortParse("2020-12-31T23:59:59")).priceList(Long.valueOf(4)).price(Double.valueOf(38.95)).build())
 		.toList();
 	
 	// This list merges both (each item contains the request and response to that request)
